requirements.txt
Django==4.2.10 
gunicorn==20.1.0 
whitenoise==6.5.0 
psycopg2-binary==2.9.9 
python-dotenv==1.0.0 
dj-database-url==2.0.0 
